+++ lib/pyseq/__init__.py
@@ -48,6 +48,6 @@ Docs and latest version available for download at
 """
 
 __author__ = "Ryan Galloway"
-__version__ = "0.9.1"
+__version__ = "0.9.2"
 
 from .seq import *